MIgrate commands to Console
