+++ src/Drupal/Migrate/ValidateMigrationId.php
@@ -7,7 +7,7 @@ namespace Drush\Drupal\Migrate;
 use Attribute;
 use Drush\Attributes\NoArgumentsBase;
 
-#[Attribute(Attribute::TARGET_METHOD)]
+#[Attribute(Attribute::TARGET_METHOD | Attribute::TARGET_CLASS)]
 class ValidateMigrationId extends NoArgumentsBase
 {
     protected const NAME = 'validate_migration_id';
