+++ tests/integration/MigrateRunnerTest.php
@@ -20,7 +20,7 @@ class MigrateRunnerTest extends UnishIntegrationTestCase
     protected function setUp(): void
     {
         parent::setUp();
-        $this->drush('pm:install', ['migrate', 'node', 'woot']);
+        $this->drush('pm:install', ['migrate', 'node', 'woot'], ['yes' => null]);;
     }
 
     /**
